Fix round for float value when input out of long range

+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/MathFunctions.java
@@ -1136,7 +1136,7 @@ public final class MathFunctions
         catch (ArithmeticException e) {
             // Use BigDecimal if the value is out of the range of long.
             BigDecimal bigDecimal = new BigDecimal(numInFloat);
-            return floatToRawIntBits(bigDecimal.setScale((int) decimals, HALF_UP).longValue());
+            return floatToRawIntBits(bigDecimal.setScale((int) decimals, HALF_UP).floatValue());
         }
     }
